requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
jeepney==0.9.0
loguru==0.6.0
notify_py==0.3.43
packaging==25.0
pillow==12.0.0
